Move reused v2 addon to the scenarios file
